* Make `vg index --store-alignments` efficient enough for WGS GAMs by wandering through the [RocksDB options jungle](https://github.com/facebook/rocksdb/wiki/RocksDB-Tuning-Guide) to tune buffer sizes, concurrency, and write amplification.
* We no longer perform a total compaction of the RocksDB database upon initial completion of the load. The tradeoff is more work to do during subsequent reads from the index, as RocksDB needs to search within a number of sorted tables instead of just one. We do however keep this number small using background compactions while the load is running, so the practical impact should be minimal compared to the frustration of waiting for the single-threaded total compaction (which can still be done using `vg index --compact`).
* `vg index --store-alignments` would crash on any 'unmapped' read. For now I've avoided the crash by making it ignore/discard such reads and printing a warning (as a temporary FIXME measure). As a next step, we should decide how these should be keyed in the database.

This is a start- lots more to do!