@mlin 2016-11-29
@ekg the load performance remains quite dependent on storage bandwidth. make sure to build the rocksdb index on a locally attached scratch SSD if possible rather than NAS. once built of course it can be copied and accessed from NAS.