@ekg 2016-11-30
@mlin, this really worked well.

Before, on a yeast data set:

```
eg10@vr-4-1-16 [17:37:26] [/lustre/scratch115/realdata/mdt3/projects/graphs/yeast/erik] 
-> % time vg index -N -d Y12+SK1.gam.index aln.gam
vg index -N -d Y12+SK1.gam.index aln.gam  6673.46s user 3173.51s system 254% cpu 1:04:22.31 total
```

After:

```
eg10@vr-4-1-16 [18:41:50] [/lustre/scratch115/realdata/mdt3/projects/graphs/yeast/erik] 
-> % time vg index -N -d Y12+SK1.gam.index.1 aln.gam
vg index -N -d Y12+SK1.gam.index.1 aln.gam  3330.13s user 1416.78s system 838% cpu 9:26.16 total
```

6 times faster!